packages/creature-generator/requirements.txt
flask
Pillow
requests
dotenv
